+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fixed decimals for syncing percentage ([PR 3106](https://github.com/input-output-hk/daedalus/pull/3106))
 
+### Chores
+
+- Updated `@cardano-foundation/ledgerjs-hw-app-cardano` to version `6.0.0` ([PR 3093](https://github.com/input-output-hk/daedalus/pull/3093))
+
 ## 5.2.0
 
 ### Features